Adopt shared database server targets for Geode
