+++ instrumentation/geode-1.4/javaagent-unit-tests/build.gradle.kts
@@ -3,5 +3,6 @@ plugins {
 }
 
 dependencies {
+  testImplementation(project(":instrumentation-api-incubator"))
   testImplementation(project(":instrumentation:geode-1.4:javaagent"))
 }
